# gen/http/openapi.yaml
swagger: "2.0"
info:
    title: GRPC-REST api for users, orders, products
    description: This api manages CRUD operations
    version: 0.0.1
host: localhost:8080
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
    /api/v1/prodcut/{ProductSKU}:
        post:
            tags:
                - product
            summary: addproduct product
            operationId: product#addproduct
            parameters:
                - name: ProductSKU
                  in: path
                  description: Product SKU number
                  required: true
                  type: string
                - name: AddproductRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/ProductAddproductRequestBody'
                    required:
                        - ProductName
                        - ProductDescription
                        - ProductPrice
            responses:
                "201":
                    description: Created response.
            schemes:
                - http
    /api/v1/product:
        get:
            tags:
                - product
            summary: showproduct product
            operationId: product#showproduct
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/ProductProductManagementResponseCollection'
            schemes:
                - http
    /api/v1/product/{ProductSKU}:
        get:
            tags:
                - product
            summary: getproduct product
            operationId: product#getproduct
            parameters:
                - name: ProductSKU
                  in: path
                  description: Product User
                  required: true
                  type: string
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/ProductGetproductResponseBody'
            schemes:
                - http
    /openapi.json:
        get:
            tags:
                - product
            summary: Download ./gen/http/openapi.json
            operationId: product#/openapi.json
            responses:
                "200":
                    description: File downloaded
                    schema:
                        type: file
            schemes:
                - http
definitions:
    ProductAddproductRequestBody:
        title: ProductAddproductRequestBody
        type: object
        properties:
            ProductDescription:
                type: string
                description: Product description
                example: Quia ab cupiditate aut tempore voluptatibus in.
            ProductName:
                type: string
                description: Product Name
                example: Quo in sint et harum velit.
            ProductPrice:
                type: number
                description: Product Price
                example: 0.5805810906976397
                format: double
        example:
            ProductDescription: Aspernatur error dolorem harum quidem.
            ProductName: Nostrum aspernatur aut velit veniam iste.
            ProductPrice: 0.667296122354781
        required:
            - ProductName
            - ProductDescription
            - ProductPrice
    ProductGetproductResponseBody:
        title: 'Mediatype identifier: application/vnd.product; view=default'
        type: object
        properties:
            ProductDescription:
                type: string
                description: Description of the Product.
                example: Organic orange of 234g
            ProductName:
                type: string
                description: Name of the Product.
                example: Orange
            ProductPrice:
                type: number
                description: Price of the Product.
                example: 6.85
                format: double
            ProductSKU:
                type: string
                description: SKU is the unique id of the Product.
                example: "12356890"
        description: GetproductResponseBody result type (default view)
        example:
            ProductDescription: Organic orange of 234g
            ProductName: Orange
            ProductPrice: 6.85
            ProductSKU: "12356890"
        required:
            - ProductSKU
            - ProductName
            - ProductDescription
            - ProductPrice
    ProductManagementResponse:
        title: 'Mediatype identifier: application/vnd.product; view=default'
        type: object
        properties:
            ProductDescription:
                type: string
                description: Description of the Product.
                example: Organic orange of 234g
            ProductName:
                type: string
                description: Name of the Product.
                example: Orange
            ProductPrice:
                type: number
                description: Price of the Product.
                example: 6.85
                format: double
            ProductSKU:
                type: string
                description: SKU is the unique id of the Product.
                example: "12356890"
        description: A ProductManagement type describes a product of shop. (default view)
        example:
            ProductDescription: Organic orange of 234g
            ProductName: Orange
            ProductPrice: 6.85
            ProductSKU: "12356890"
        required:
            - ProductSKU
            - ProductName
            - ProductDescription
            - ProductPrice
    ProductProductManagementResponseCollection:
        title: 'Mediatype identifier: application/vnd.product; type=collection; view=default'
        type: array
        items:
            $ref: '#/definitions/ProductManagementResponse'
        description: ShowproductResponseBody is the result type for an array of ProductManagementResponse (default view)
        example:
            - ProductDescription: Organic orange of 234g
              ProductName: Orange
              ProductPrice: 6.85
              ProductSKU: "12356890"
            - ProductDescription: Organic orange of 234g
              ProductName: Orange
              ProductPrice: 6.85
              ProductSKU: "12356890"
            - ProductDescription: Organic orange of 234g
              ProductName: Orange
              ProductPrice: 6.85
              ProductSKU: "12356890"
            - ProductDescription: Organic orange of 234g
              ProductName: Orange
              ProductPrice: 6.85
              ProductSKU: "12356890"